By user systemd service/socket
Instead of having a socket for a static and predefined user http, this
patch allow to easily setup multiple sockets with your needed users.

This is particulary useful when you want have socket for dedicated user.

You can create an http socket (in /run/fcgiwrap-http.sock) with:
$ systemctl enable fcgiwrap@http.socket

and if you need a socket for munin (in /run/fcgiwrap-munin.sock):
$ systemctl enable fcgiwrap@munin.socket