Attempt to correct usage of size_t
size_t is assumed to be synonymous with the types int, ptrdiff_t,
ssize_t and socklen_t. It is not specified whether variables of these
types can hold values of type size_t and vice versa. However, it could
be checked at compile time whether these assumptions hold on a
particular system.