Add size_t workaround for FCGI_fread and FCGI_fwrite
FCGI_fread and FCGI_fwrite can return (size_t)EOF. (size_t)EOF is not a
valid value, because EOF is defined as negative and size_t as unsigned.
FCGI_write can also return -1 / size which is also not a valid value. As
a workaround wrapper functions check whether FCGI_fread and FCGI_fwrite
return these invalid value and report an error. However, it could be
that -1 / size and (size_t)EOF are valid size_t values. There is no way
to distinguish this and it is better to fail than to proceed with an
invalid value.