Makefile.in: Use LIBS from configure rather than specifying -lfcgi manually.
fcgiwrap uses libfcgi, and libfcgi uses libm.  This is an issue when
using static libs, because -lm does not get properly added to the link
flags in this case.  A solution is to provide LIBS=-lm to configure
when needed.  However, fcgiwrap does not use Automake and its
Makefile.in ignores LIBS from configure.  Fix that.

Signed-off-by: Thomas Claveirole <thomas.claveirole@green-communications.fr>